% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{apk-class}
\alias{apk-class}
\title{Class: apk model}
\description{
To create an apk object, use \link[aPack]{apk}. See also this function for mode details.
}
\section{Slots}{

\describe{
\item{\code{call}}{Object of class \code{"language"}. User call reminder.}
}}

