% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{apk}
\alias{apk}
\title{Create an Object of class \code{"apk"}}
\usage{
apk(foo, ...)
}
\arguments{
\item{foo}{Not used yet.}

\item{...}{Not used yet.}
}
\description{
Creator function for objects of class \code{"apk"}.
}
