% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict}
\alias{predict}
\alias{predict,apk-method}
\title{Prediction Method for the apk Class}
\usage{
predict(object, ...)

\S4method{predict}{apk}(object, bar, ...)
}
\arguments{
\item{object}{An object to predict from.}

\item{...}{Further arguments for methods.}

\item{bar}{Not used yet.}
}
\description{
Prediction Method for the apk Class
}
\examples{
myApk <- apk()
predict(myApk)
}
